refactor: handle dashboard filters in query variant controller
